Fix for #4935 Runs upgrade test with alpha,beta and dynamic clients
