+++ test/upgrade/v1alpha1/service_preupgrade_test.go
@@ -1,0 +1,34 @@
+// +build preupgrade
+
+/*
+Copyright 2018 The Knative Authors
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+package v1alpha1
+
+import (
+	"knative.dev/serving/test/upgrade"
+	"testing"
+)
+
+func TestRunLatestServicePreUpgrade(t *testing.T) {
+	t.Parallel()
+	upgrade.CreateService(t, "v1alpha1")
+}
+
+func TestRunLatestServicePreUpgradeAndScaleToZero(t *testing.T) {
+	t.Parallel()
+	upgrade.CreateServiceAndScaleToZero(t, "v1alpha1")
+}
